Web_GUI/requirements.txt
opencv-python==4.11.0.86
numpy==2.2.1
mjpeg-streamer
python-can==4.5.0
cantools==40.2.1
nicegui==2.18.0
plotly
pyserial
plotly-resampler